refactor(templates): create a folder to all templates
affects: @crave/farmblocks-dev-scaffold
