+++ packages/scaffold/src/templates/package.js
