+++ packages/scaffold/src/templates/readme.js
